refactor: moves validation logic into a separate component
