+++ src/app/components/password-strength/password-strength.component.html
@@ -12,10 +12,9 @@
         [type]="isVisible ? 'text' : 'password'"
         class="font-medium text-gray-700 outline-none pl-3 py-2 w-4/5"
         placeholder="Type password here..."
-        (input)="validatePassword()"
       />
       <button
-        (click)="toggleVisible()"
+        (click)="togglePasswordVisible()"
         class="absolute right-0 h-full font-medium text-gray-500 w-1/5 rounded-xl"
       >
         {{ isVisible ? "Hide" : "Show" }}
