+++ src/app/components/password-strength/password-strength.component.ts
@@ -1,11 +1,6 @@
 import { Component } from '@angular/core';
 import { passwordStrengthConfig } from './password-strength.config';
-
-interface Types {
-  hasDigits?: boolean;
-  hasLetters?: boolean;
-  hasSymbols?: boolean;
-}
+import { PasswordValidatorService } from 'src/app/services/password-validator.service';
 
 type passwordStrengthType = 'empty' | 'easy' | 'short' | 'medium' | 'strong';
 
@@ -14,54 +9,19 @@ type passwordStrengthType = 'empty' | 'easy' | 'short' | 'medium' | 'strong';
   templateUrl: './password-strength.component.html',
 })
 export class PasswordStrengthComponent {
-  password = '';
+  private _password = '';
   passwordStrength: passwordStrengthType = 'empty';
   isVisible = false;
 
-  private typesCount(password: string): number {
-    const types: Types = {};
+  constructor(private passwordService: PasswordValidatorService) {}
 
-    for (let char of password) {
-      if (this.isDigit(char)) {
-        types.hasDigits = true;
-      } else if (this.isLetter(char)) {
-        types.hasLetters = true;
-      } else {
-        types.hasSymbols = true;
-      }
-    }
-
-    const typesLength = Object.keys(types).length;
-    return typesLength;
+  set password(value: string) {
+    this._password = value;
+    this.passwordStrength = this.passwordService.validate(this._password);
   }
 
-  toggleVisible(): void {
+  togglePasswordVisible(): void {
     this.isVisible = !this.isVisible;
-  }
-
-  validatePassword() {
-    const password = this.password.trim();
-    const typesCount = this.typesCount(password);
-
-    if (password === '') {
-      this.passwordStrength = 'empty';
-    } else if (password.length < 8) {
-      this.passwordStrength = 'short';
-    } else if (typesCount === 1) {
-      this.passwordStrength = 'easy';
-    } else if (typesCount === 2) {
-      this.passwordStrength = 'medium';
-    } else {
-      this.passwordStrength = 'strong';
-    }
-  }
-
-  isDigit(char: string): boolean {
-    return /\d/.test(char);
-  }
-
-  isLetter(char: string): boolean {
-    return /[a-zA-Z]/.test(char);
   }
 
   getColorClass(position: 'left' | 'middle' | 'right'): string {
